adds commonStore and login header component
