+++ config/webpack.config.js
@@ -15,9 +15,9 @@ const configGenerator = (options) => {
       facilities: './src/js/facility-locator/facility-locator-entry.jsx',
       hca: './src/js/hca/hca-entry.jsx',
       messaging: './src/js/messaging/messaging-entry.jsx',
-      'no-react': './src/js/no-react-entry.js',
-      'user-profile': './src/js/user-profile/user-profile-entry.jsx',
-       rx: './src/js/rx/rx-entry.jsx',
+      rx: './src/js/rx/rx-entry.jsx',
+      'no-react': ['./src/js/no-react-entry.js', './src/js/login/login-entry.jsx'],
+      'user-profile': ['./src/js/user-profile/user-profile-entry.jsx', './src/js/login/login-entry.jsx'],
     },
     output: {
       path: path.join(__dirname, `../build/${options.buildtype}/generated`),
