+++ content/includes/header.html
@@ -132,6 +132,7 @@
       </div>
       <!-- htmllint attr-bans="$previous" id-no-dup="$previous" -->
     {% endif %}
+    <div class="text-right columns" id="login-root"></div>
   </div>
 </div>
