+++ src/js/common/actions/index.js
@@ -1,6 +1,7 @@
 export const UPDATE_LOGGEDIN_STATUS = 'UPDATE_LOGGEDIN_STATUS';
 export const UPDATE_LOGIN_URL = 'UPDATE_LOGIN_URL';
 export const UPDATE_PROFILE_FIELD = 'UPDATE_PROFILE_FIELD';
+export const LOG_OUT = 'LOG_OUT';
 
 export function updateLoggedInStatus(value) {
   return {
@@ -21,5 +22,11 @@ export function updateProfileField(propertyPath, value) {
     type: UPDATE_PROFILE_FIELD,
     propertyPath,
     value
+  };
+}
+
+export function logOut() {
+  return {
+    type: LOG_OUT
   };
 }
