+++ src/js/common/reducers/index.js
