+++ src/js/common/reducers/login.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import lodashDeep from 'lodash-deep';
 
-import { UPDATE_LOGGEDIN_STATUS, UPDATE_LOGIN_URL } from '../actions';
+import { UPDATE_LOGGEDIN_STATUS, UPDATE_LOGIN_URL, LOG_OUT } from '../actions';
 
 // Add deep object manipulation routines to lodash.
 _.mixin(lodashDeep);
@@ -23,6 +23,10 @@ function loginStuff(state = initialState, action) {
     case UPDATE_LOGIN_URL:
       newState = Object.assign({}, state);
       _.set(newState, 'loginUrl', action.value);
+      return newState;
+
+    case LOG_OUT:
+      newState = initialState;
       return newState;
 
     default:
