+++ src/js/common/reducers/profile.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import lodashDeep from 'lodash-deep';
 
-import { UPDATE_PROFILE_FIELD } from '../actions';
+import { UPDATE_PROFILE_FIELD, LOG_OUT } from '../actions';
 
 // Add deep object manipulation routines to lodash.
 _.mixin(lodashDeep);
@@ -24,6 +24,11 @@ function profileInformation(state = initialState, action) {
     case UPDATE_PROFILE_FIELD: {
       newState = Object.assign({}, state);
       _.set(newState, action.propertyPath, action.value);
+      return newState;
+    }
+
+    case LOG_OUT: {
+      newState = initialState;
       return newState;
     }
