+++ src/js/login/login-entry.jsx
@@ -1,16 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { Router, browserHistory } from 'react-router';
 import { Provider } from 'react-redux';
 
 import initReact from '../common/init-react';
-import routes from './routes';
 import { commonStore } from '../common/store';
 
+import Main from './containers/Main';
+
 require('../common');  // Bring in the common javascript.
-require('../../sass/rx/rx.scss');
-require('../../sass/user-profile.scss');
+require('../../sass/login.scss');
 
 function init() {
   /*
@@ -20,18 +19,12 @@ function init() {
    * Plan is to make this trigger a sort when the query
    * parameter is `sortby`.
    */
-  const handleChangedURL = (event) => {
-    // Here so eslint doesn't tell us about an unused variable.
-    return event;
-  };
-  browserHistory.listen(handleChangedURL);
-  // End URL listening
 
   ReactDOM.render((
     <Provider store={commonStore}>
-      <Router history={browserHistory} routes={routes}/>
+      <Main/>
     </Provider>
-    ), document.getElementById('react-root'));
+    ), document.getElementById('login-root'));
 }
 
 initReact(init);
