+++ src/sass/login.scss
@@ -1,0 +1,7 @@
+// Variables
+
+@import 'style';
+
+.signned-in-link {
+  background: transparent;
+};
